refactor: reorganize add post dialog components and related API logic
