+++ src/entity/post/api/addPost.ts
@@ -1,0 +1,11 @@
+import { NewPost } from "../model/types"
+
+export const addPost = async (post: NewPost) => {
+  const response = await fetch("/api/posts/add", {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(post),
+  })
+  const data = await response.json()
+  return data
+}
